requirements.txt
django
gunicorn
django-heroku
psycopg2
django-tinymce
django-crispy-forms
